Refactored to separate resources into their own files
